requirements.txt
langgraph
langchain_aws
boto3
kubernetes
fastapi
uvicorn
python-dotenv